requirements.txt
streamlit
transformers
torch
google-generativeai
numpy
scikit-learn